Major reworking of the source to remove the need to generate mocks for Android SDK classes at compile time.
This addresses:           

- Building your tests for platform X (e.g. 2.2) and running on platform Y where X < Y (e.g. Y = 2.1).  This sounds a bit strange,
but there actually is demand for this.
- Building your tests for platform 2.0 and running on platform 2.2 in case of changes to the SDK classes that will cause verify errors
at runtime.  Although Android Mock tries to rely only on public (and hence for the SDK, stable) parts of the classes being mocked,
some class internal changes do sometimes get relied upon.  Greater reliance on internals will happen in the future (allowing mocking
of package-private members for instance).
- Removes the need for the genmocks. package prefix in front of non-SDK classes.  This convention was introduced to avoid conflicts
with Android when placing classes in an SDK package (which causes compilation errors).  Doing this prevented the mocking of
package-private members.  Now that the restriction is lifted, a future change can rather simply make it possible to enable
package-private member mocking.


This change:
- Android SDK class mocks are now pre-built.
- They are picked up at compile time by the system and then added to your Test APK.
- At runtime, Android Mock resolves which version to use.